fix hapus nota, mulai sj

+++ database/migrations/2021_09_21_134100_spk_produk.php
@@ -31,12 +31,12 @@ class SpkProduk extends Migration
             $table->foreignId('spk_id')->constrained()->onDelete('cascade')->onUpdate('cascade');
             $table->foreignId('produk_id');
             $table->string('ktrg')->nullable();
-            $table->integer('jumlah');
-            $table->integer('deviasi_jml')->nullable()->default(0);
-            $table->integer('jml_t')->nullable()->default(0);
-            $table->integer('jml_selesai')->nullable()->default(0);
-            $table->integer('jml_blm_sls')->nullable()->default(0);
-            $table->integer('jml_sdh_nota')->nullable()->default(0);
+            $table->mediumInteger('jumlah');
+            $table->mediumInteger('deviasi_jml')->nullable()->default(0);
+            $table->mediumInteger('jml_t')->nullable()->default(0);
+            $table->mediumInteger('jml_selesai')->nullable()->default(0);
+            $table->mediumInteger('jml_blm_sls')->nullable()->default(0);
+            $table->mediumInteger('jml_sdh_nota')->nullable()->default(0);
             $table->integer('harga');
             $table->integer('koreksi_harga')->nullable();
             $table->string('status', 20)->nullable(); // Status yang berkaitan dengan sudah selesai di produksi atau belum
